+++ src/components/Footer.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import '../assets/styles/footer.css';
+
+export default function Footer() {
+    return (
+        <footer>
+            <div>
+                <p>Wszelkie prawa zastrzeżone</p>
+                <br />
+                <h3>Marek Kurańda</h3>
+                <h4>Yummy &copy; 2023</h4>
+                <br />
+                <h5>
+                    <a target="_blank" rel="noreferrer" href="https://icons8.com/icon/RxvLC54xtSnZ/yummy">Yummy</a>
+                    &nbsp;icon by&nbsp;
+                    <a target="_blank" rel="noreferrer" href="https://icons8.com">Icons8</a>
+                </h5>
+            </div>
+        </footer>
+    );
+}