+++ src/pages/Main.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import MealCategory, { MealCategoryProps } from '../components/MealCategory';
 import Button from '../components/ui/Button';
+import Footer from '../components/Footer';
 
 import '../assets/styles/header.css';
 import '../assets/styles/index-media.css';
@@ -73,20 +74,7 @@ export default function Main() {
                 </div>
             </main>
             
-            <footer>
-                <div>
-                    <p>Wszelkie prawa zastrzeżone</p>
-                    <br />
-                    <h3>Marek Kurańda</h3>
-                    <h4>Yummy &copy; 2023</h4>
-                    <br />
-                    <h5>
-                        <a target="_blank" rel="noreferrer" href="https://icons8.com/icon/RxvLC54xtSnZ/yummy">Yummy</a>
-                        &nbsp;icon by&nbsp;
-                        <a target="_blank" rel="noreferrer" href="https://icons8.com">Icons8</a>
-                    </h5>
-                </div>
-            </footer>
+            <Footer />
         </>
     );
 }