+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-02-14
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Added `Footer component`.
+- [Marek Kurańda](https://github.com/mjkuranda): Added `result` subpage.
+
 ## [0.2.0] - 2023-02-11
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Added ESLint.